Restructure utilities package
Move CLI utility functions into their own directory, create a directory for RPC relating utilities
