+++ src/utilities/__init__.py
@@ -1,9 +1,1 @@
-from .assets import DiscordAssets
-from .logger import Logger
-from .database import (
-    get_database,
-    get_game_version,
-    get_player_region,
-    get_player_union_level,
-)
-from .presence import Presence
+
