+++ src/utilities/cli/__init__.py
@@ -1,0 +1,11 @@
+from .output import indent, print_divider
+from .errors import fatal_error
+from .input import (
+    get_input,
+    get_boolean_input,
+    get_startup_preference,
+    get_shortcut_preference,
+    get_database_access_preference,
+    get_rich_presence_install_location,
+    get_wuwa_install_location,
+)
