+++ src/utilities/rpc/database.py
@@ -1,7 +1,7 @@
 from config import Config
 from sqlite3 import Connection, connect
 from json import loads
-from src.utilities import Logger
+from src.utilities.rpc import Logger
 
 
 def get_database() -> Connection:
