+++ src/utilities/rpc/presence.py
@@ -3,12 +3,13 @@ from psutil import Process, NoSuchProcess, pids
 from sqlite3 import Connection
 from pypresence import Presence as PyPresence
 from config import Config
-from src.utilities import Logger, DiscordAssets
-from src.utilities import (
+from src.utilities.rpc import (
     get_database,
     get_player_region,
     get_player_union_level,
     get_game_version,
+    DiscordAssets,
+    Logger,
 )
